Update transform tests
